requirements-dev.txt
-r requirements.txt
Fabric==1.6.0
paramiko==1.10.1
pycrypto==2.6
django-webtest==1.7.5
webassets==0.8
WebOb==1.2.3
waitress==0.8.7
beautifulsoup4==4.3.2
factory-boy==2.2.1
